Fix "on behalf of" display in outlook
If you use a subdomain in mailgun, as described in their official docs, the from address in Outlook is displayed strange, which could confuse users. You can read about it here https://help.mailgun.com/hc/en-us/articles/360012491394-Why-Do-I-See-On-Behalf-Of-in-My-Email-
The problem can be solved when from and sender headers are the same. The solution was found here https://stackoverflow.com/a/45445015/616067